final submit

+++ include/PWM_ISR.h
@@ -22,7 +22,7 @@ typedef struct _FeedbackVarBuf{
 extern FeedbackVarBuf feedbackVarBuf;
 extern Uint16 test_data[160];
 extern Uint16 al[160];
-extern Uint16 alb[160];
+extern int alb[160];
 extern Uint16 alc[160];
 extern Uint16 kal[160];
 #endif
